Removed assetId from CodeEditor, and updated dropping file and pasting code
