+++ webstack/libs/frontend/src/lib/utils/setupAppForFiles.ts
@@ -23,6 +23,7 @@ import {
   isGIF,
   isCode,
   isFileURL,
+  stringContainsCode,
 } from '@sage3/shared';
 import { apiUrls } from '@sage3/frontend';
 import { ExtraImageType, ExtraPDFType, FileEntry, User } from '@sage3/shared/types';
@@ -99,6 +100,19 @@ export async function setupAppForFile(
       }
     }
   } else if (isCode(file.type)) {
+    // Look for the file in the asset store
+    const localurl = apiUrls.assets.getAssetById(file.filename);
+    // Get the content of the file
+    const response = await fetch(localurl, {
+      headers: {
+        'Content-Type': 'text/plain',
+        Accept: 'text/plain',
+      },
+    });
+    // Get the content of the file
+    const text = await response.text();
+    // Get Lanauge
+    const lang = stringContainsCode(text);
     return {
       title: 'CodeEditor',
       roomId: roomId,
@@ -107,7 +121,7 @@ export async function setupAppForFile(
       size: { width: 850, height: 400, depth: 0 },
       rotation: { x: 0, y: 0, z: 0 },
       type: 'CodeEditor',
-      state: { ...(initialValues['CodeEditor'] as AppState), assetid: file.id, content: '', language: '' },
+      state: { ...(initialValues['CodeEditor'] as AppState), content: text, language: lang },
       raised: true,
       dragging: false,
       pinned: false,
